- turned off max-classes-per-file and updated type files - updated serializers to have toJSON suffix - moved access token logic to service - refactored AccountAuthMiddleware.ensureAccess to accessAuthMiddleware - fixed DatadogLogger
