+++ .eslintrc
@@ -91,7 +91,8 @@
           }
         ],
         "import/extensions": "off",
-        "import/prefer-default-export": "off"
+        "import/prefer-default-export": "off",
+        "max-classes-per-file": "off"
       }
     },
     {
