+++ src/apps/backend/modules/access-token/access-token-service.ts
@@ -1,6 +1,12 @@
-import AccessTokenWriter from './internal/access-token-writer';
+import jsonwebtoken from 'jsonwebtoken';
+
+import AccountService from '../account/account-service';
+import { ConfigService } from '../config';
+
 import {
   AccessToken,
+  AccessTokenExpiredError,
+  AccessTokenInvalidError,
   AccessTokenPayload,
   CreateAccessTokenParams,
   VerifyAccessTokenParams,
@@ -10,12 +16,54 @@ export default class AccessTokenService {
   public static async createAccessToken(
     params: CreateAccessTokenParams,
   ): Promise<AccessToken> {
-    return AccessTokenWriter.createAccessToken(params);
+    const account = await AccountService.getAccountByUsernamePassword({
+      password: params.password,
+      username: params.username,
+    });
+
+    const jwtSigningKey = ConfigService.getValue<string>('accounts.tokenKey');
+    const jwtExpiry = ConfigService.getValue<string>('accounts.tokenExpiry');
+
+    const jwtToken = jsonwebtoken.sign(
+      { accountId: account.id },
+      jwtSigningKey,
+      { expiresIn: jwtExpiry },
+    );
+
+    const accessToken = new AccessToken();
+    accessToken.accountId = account.id;
+    accessToken.token = jwtToken;
+
+    const jwtTokenDecoded = jsonwebtoken.decode(jwtToken) as jsonwebtoken.JwtPayload;
+    accessToken.expiresAt = new Date(jwtTokenDecoded.exp * 1000);
+
+    return accessToken;
   }
 
   public static verifyAccessToken(
     params: VerifyAccessTokenParams,
   ): AccessTokenPayload {
-    return AccessTokenWriter.verifyAccessToken(params);
+    const { token } = params;
+    let verifiedToken: jsonwebtoken.JwtPayload;
+
+    try {
+      const jwtSigningKey = ConfigService.getValue<string>('accounts.tokenKey');
+
+      verifiedToken = jsonwebtoken.verify(
+        token,
+        jwtSigningKey,
+        { ignoreExpiration: true },
+      ) as jsonwebtoken.JwtPayload;
+    } catch (e) {
+      throw new AccessTokenInvalidError();
+    }
+
+    if (verifiedToken.exp * 1000 < Date.now()) {
+      throw new AccessTokenExpiredError();
+    }
+
+    return {
+      accountId: verifiedToken.accountId as string,
+    };
   }
 }
