+++ src/apps/backend/modules/access-token/index.ts
@@ -1,4 +1,4 @@
 export { default as AccessTokenServer } from './rest-api/access-token-server';
-export { default as AccountAuthMiddleware } from './rest-api/account-auth-middleware';
 export { default as AccessTokenService } from './access-token-service';
+export * from './rest-api/access-auth-middleware';
 export * from './types';
