+++ src/apps/backend/modules/access-token/rest-api/access-token-serializer.ts
@@ -1,6 +1,6 @@
 import { AccessToken } from '../types';
 
-export const serializeAccessToken = (accessToken: AccessToken): unknown => ({
+export const serializeAccessTokenAsJSON = (accessToken: AccessToken): unknown => ({
   accountId: accessToken.accountId,
   expiresAt: accessToken.expiresAt.toUTCString(),
   token: accessToken.token,
