+++ src/apps/backend/modules/access-token/types.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 import { AppError } from '../error';
 import { HttpStatusCodes } from '../http';
