+++ src/apps/backend/modules/account/rest-api/account-serializer.ts
@@ -1,6 +1,6 @@
 import { Account } from '../types';
 
-export const serializeAccount = (account: Account): unknown => ({
+export const serializeAccountAsJSON = (account: Account): unknown => ({
   id: account.id,
   username: account.username,
 });
