+++ src/apps/backend/modules/communication/types.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-classes-per-file */
 import { AppError } from '../error';
 import { HttpStatusCodes } from '../http';
