+++ src/apps/backend/modules/logger/internals/datadog-logger.ts
@@ -1,21 +1,21 @@
 import * as winston from 'winston';
 
-import ConfigService from '../../config/config-service';
+import { ConfigService } from '../../config';
 
 import WinstonLogger from './winston-logger';
 
 export default class DatadogLogger extends WinstonLogger {
   constructor() {
-    const DATADOG_API_KEY = ConfigService.getStringValue('datadog.api_key');
-    const APP_NAME = ConfigService.getStringValue('datadog.app_name');
+    const DD_REGION = ConfigService.getValue<string>('datadog.region');
+    const DD_API_KEY = ConfigService.getValue<string>('datadog.api_key');
+    const DD_APP_NAME = ConfigService.getValue<string>('datadog.app_name');
 
-    const httpTransportOptions = {
-      host: 'http-intake.logs.us5.datadoghq.com',
-      path: `/api/v2/logs?dd-api-key=${DATADOG_API_KEY}&ddsource=nodejs&service=${APP_NAME}`,
+    const transport = new winston.transports.Http({
+      host: `http-intake.logs.${DD_REGION}.datadoghq.com`,
+      path: `/api/v2/logs?dd-api-key=${DD_API_KEY}&ddsource=nodejs&service=${DD_APP_NAME}`,
       ssl: true,
-    };
+    });
 
-    const transport = new winston.transports.Http(httpTransportOptions);
     super(transport);
   }
 }
