+++ src/apps/backend/modules/task/rest-api/task-router.ts
@@ -1,4 +1,4 @@
-import { AccountAuthMiddleware } from '../../access-token';
+import { accessAuthMiddleware } from '../../access-token';
 import { ApplicationRouter } from '../../application';
 
 import * as TaskController from './task-controller';
@@ -7,7 +7,7 @@ export default class TaskRouter extends ApplicationRouter {
   configure(): void {
     const { router } = this;
 
-    router.use(AccountAuthMiddleware.ensureAccess);
+    router.use(accessAuthMiddleware);
 
     router.post('/', TaskController.createTask);
     router.get('/', TaskController.getTasks);
